moving get_page_state_url back to templatetags
