+++ cfgov/v1/templatetags/share.py
@@ -1,9 +1,25 @@
 from django import template
-
-from v1.models import CFGOVUserPagePermissionsProxy
+from django.conf import settings
+from urlparse import urlparse
 from wagtail.wagtailcore.models import Page
 
+
 register = template.Library()
+
+
+@register.assignment_tag(takes_context=True)
+def get_page_state_url(context, page):
+    page = Page.objects.get(id=page.id).specific
+    if not page.live and not page.shared:
+        return None
+    url = page.url
+    if url is None:  # If page is not aligned to a site root return None
+        return None
+    page_hostname = urlparse(url).hostname
+    staging_hostname = settings.STAGING_HOSTNAME
+    if not page.live and page.shared and staging_hostname != page_hostname:
+        url = url.replace(page_hostname, staging_hostname)
+    return url
 
 
 @register.filter
@@ -20,6 +36,9 @@ def is_shared(page):
 def v1page_permissions(context, page):
     page = page.specific
     if 'user_page_permissions' not in context:
-        context['user_page_permissions'] = CFGOVUserPagePermissionsProxy(context['request'].user)
+        from v1.models import CFGOVUserPagePermissionsProxy
+        context['user_page_permissions'] = CFGOVUserPagePermissionsProxy(
+            context['request'].user
+        )
 
     return context['user_page_permissions'].for_page(page)
